fix: EABI32 mappers for 5+ arg functions, prxgen SHT_REL fix, i8/i9 bridges
PSP Allegrex uses EABI32 (args 5-8 in $t0-$t3) but LLVM generates O32
(args 5+ on stack). This mismatch caused garbage in $t0 for 5+ arg
functions, breaking sceKernelStartModule, semaphores, event flags, HTTP
connections, and other critical APIs.

Changes:
- eabi.rs: Add i8 (8-arg) and i9 (9-arg) EABI bridge assembly functions
- kernel/mod.rs: sceKernelStartModule (i5), sceKernelLoadModuleByID (i5)
- kernel/thread.rs: sceKernelCreateSema, WaitEventFlag, WaitEventFlagCB,
  CreateLwMutex, CreateThread (i5/i6)
- net.rs: 16 functions with i5-i9 mappers (sceNetInit, HTTP, adhoc, etc.)
  Also widen u16 port params to u32 (kernel reads full 32-bit registers)
- font.rs: sceFontGetCharGlyphImage_Clip, ShadowGlyphImage_Clip (i5)
- macros.rs: Document EABI transmute behavior for u16/u64 edge cases
- mpeg.rs: Disable import stubs (cfg(any)), add dummy mpeg_stubs module
  for function-pointer dispatch (sceMpegVsh_library import stubs cause
  freeze on ARK-4 CFW when kernel re-links after module load)
- prxgen.rs: NULL out non-ALLOC SHT_REL sections that caused kernel
  state corruption during module loading
- lib.rs: Fix module version byte order, syslib version (0,0)

+++ psp/src/sys/net.rs
@@ -1,4 +1,4 @@
-use crate::eabi::{i5, i6};
+use crate::eabi::{i5, i6, i7, i8, i9};
 use core::ffi::c_void;
 
 #[repr(C)]
@@ -14,7 +14,7 @@ psp_extern! {
     #![flags = 0x0009]
     #![version = (0x00, 0x00)]
 
-    #[psp(0x39AF39A6)]
+    #[psp(0x39AF39A6, i5)]
     /// Initialise the networking library
     ///
     /// # Parameters
@@ -276,7 +276,7 @@ psp_extern! {
     /// 0 on success, < 0 on error.
     pub fn sceNetAdhocctlGetAdhocId(id: *mut SceNetAdhocctlAdhocId) -> i32;
 
-    #[psp(0xA5C055CE)]
+    #[psp(0xA5C055CE, i6)]
     /// Connect to the Adhoc control game mode (as a host)
     ///
     /// # Parameters
@@ -583,7 +583,7 @@ psp_extern! {
         unk1: i32,
     ) -> i32;
 
-    #[psp(0xABED3790)]
+    #[psp(0xABED3790, i7)]
     /// Set a PDP packet to a destination
     ///
     /// # Parameters
@@ -602,14 +602,14 @@ psp_extern! {
     pub fn sceNetAdhocPdpSend(
         id: i32,
         dest_mac_addr: *mut u8,
-        port: u16,
+        port: u32,
         data: *mut c_void,
         len: u32,
         timeout: u32,
         nonblock: i32,
     ) -> i32;
 
-    #[psp(0xDFE53E03)]
+    #[psp(0xDFE53E03, i7)]
     /// Receive a PDP packet
     ///
     /// # Parameters
@@ -729,7 +729,7 @@ psp_extern! {
     /// 0 on success, < 0 on error.
     pub fn sceNetAdhocGameModeDeleteReplica(id: i32) -> i32;
 
-    #[psp(0x877F6D66)]
+    #[psp(0x877F6D66, i8)]
     /// Open a PTP connection
     ///
     /// # Parameters
@@ -748,9 +748,9 @@ psp_extern! {
     /// A socket ID on success, < 0 on error.
     pub fn sceNetAdhocPtpOpen(
         srcmac: *mut u8,
-        srcport: u16,
+        srcport: u32,
         destmac: *mut u8,
-        destport: u16,
+        destport: u32,
         buf_size: u32,
         delay: u32,
         count: i32,
@@ -775,7 +775,7 @@ psp_extern! {
         nonblock: i32,
     ) -> i32;
 
-    #[psp(0xE08BDAC1)]
+    #[psp(0xE08BDAC1, i7)]
     /// Wait for an incoming PTP connection
     ///
     /// # Parameters
@@ -793,7 +793,7 @@ psp_extern! {
     /// A socket ID on success, < 0 on error.
     pub fn sceNetAdhocPtpListen(
         srcmac: *mut u8,
-        srcport: u16,
+        srcport: u32,
         buf_size: u32,
         delay: u32,
         count: i32,
@@ -801,7 +801,7 @@ psp_extern! {
         unk1: i32,
     ) -> i32;
 
-    #[psp(0x9DF81198)]
+    #[psp(0x9DF81198, i5)]
     /// Accept an incoming PTP connection
     ///
     /// # Parameters
@@ -823,7 +823,7 @@ psp_extern! {
         nonblock: i32,
     ) -> i32;
 
-    #[psp(0x4DA4C788)]
+    #[psp(0x4DA4C788, i5)]
     /// Send data
     ///
     /// # Parameters
@@ -845,7 +845,7 @@ psp_extern! {
         nonblock: i32,
     ) -> i32;
 
-    #[psp(0x8BEA2B3E)]
+    #[psp(0x8BEA2B3E, i5)]
     /// Receive data
     ///
     /// # Parameters
@@ -975,7 +975,7 @@ psp_extern! {
     /// 0 on success, < 0 on error
     pub fn sceNetAdhocMatchingTerm() -> i32;
 
-    #[psp(0xCA5EDA6F)]
+    #[psp(0xCA5EDA6F, i9)]
     /// Create an Adhoc matching object
     ///
     /// # Parameters
@@ -996,7 +996,7 @@ psp_extern! {
     pub fn sceNetAdhocMatchingCreate(
         mode: AdhocMatchingMode,
         max_peers: i32,
-        port: u16,
+        port: u32,
         buf_size: i32,
         hello_delay: u32,
         ping_delay: u32,
@@ -1017,7 +1017,7 @@ psp_extern! {
     /// 0 on success, < 0 on error.
     pub fn sceNetAdhocMatchingDelete(matching_id: i32) -> i32;
 
-    #[psp(0x93EF3843)]
+    #[psp(0x93EF3843, i7)]
     /// Start a matching object
     ///
     /// # Parameters
@@ -1743,7 +1743,7 @@ psp_extern! {
     /// 0 on success, < 0 on error.
     pub fn sceHttpDeleteTemplate(templateid: i32) -> i32;
 
-    #[psp(0x8EEFD953)]
+    #[psp(0x8EEFD953, i5)]
     /// Create a http connection.
     ///
     /// # Parameters
@@ -1762,7 +1762,7 @@ psp_extern! {
         templateid: i32,
         host: *mut u8,
         unknown1: *mut u8,
-        port: u16,
+        port: u32,
         unknown2: i32,
     ) -> i32;
 
@@ -2180,7 +2180,7 @@ psp_extern! {
         err_num: *mut i32,
     ) -> i32;
 
-    #[psp(0xD70D4847)]
+    #[psp(0xD70D4847, i6)]
     pub fn sceHttpGetProxy(
         id: i32,
         activate_flag: *mut i32,
@@ -2199,13 +2199,13 @@ psp_extern! {
         cbfunc: HttpPasswordCB,
     ) -> i32;
 
-    #[psp(0xF0F46C62)]
+    #[psp(0xF0F46C62, i5)]
     pub fn sceHttpSetProxy(
         id: i32,
         activate_flag: i32,
         mode: i32,
         new_proxy_host: *const u8,
-        new_proxy_port: u16,
+        new_proxy_port: u32,
     ) -> i32;
 
     #[psp(0xC98CBBA7)]
@@ -2269,7 +2269,7 @@ psp_extern! {
     /// 0 on success, < 0 on error
     pub fn sceNetResolverDelete(rid: i32) -> i32;
 
-    #[psp(0x224C5F44)]
+    #[psp(0x224C5F44, i5)]
     /// Begin a name to address lookup
     ///
     /// # Parameters
@@ -2291,7 +2291,7 @@ psp_extern! {
         retry: i32,
     ) -> i32;
 
-    #[psp(0x629E2FB7)]
+    #[psp(0x629E2FB7, i6)]
     /// Begin a address to name lookup
     ///
     /// # Parameters
